+++ backend/Controllers/CategoriesController.cs
@@ -8,6 +8,10 @@ using backend.Dtos.GetDtos;
 using FluentValidation;
 using Microsoft.AspNetCore.Authorization;
 using System.Data;
+using backend.Dtos.Responses;
+using static System.Runtime.InteropServices.JavaScript.JSType;
+using backend.Repositories;
+using Humanizer;
 
 namespace backend.Controllers
 {
@@ -27,30 +31,30 @@ namespace backend.Controllers
 
        
         [HttpGet("{pageNumber}/{pageSize}")]
-        public async Task<ActionResult<PaginationDto<GetCategoryDto>>> GetCategories( [FromRoute] int pageNumber=1, [FromRoute] int pageSize = 4)
+        public async Task<ActionResult<APIResponse<PaginationDto<GetCategoryDto>>>> GetCategories( [FromRoute] int pageNumber=1, [FromRoute] int pageSize = 4)
         {
                 var categories = await _categoryRepository.GetAllAsync(pageSize,pageNumber);
                 var categoryDtos = _mapper.Map<PaginationDto<GetCategoryDto>>(categories);
-                return Ok(categoryDtos);
+                return Ok(new APIResponse<PaginationDto<GetCategoryDto >>(200, "",categoryDtos));
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<GetCategoryDto>> GetCategory(int id)
+        public async Task<ActionResult<APIResponse<GetCategoryDto>>> GetCategory(int id)
         {
             var category = await _categoryRepository.GetbyIdAsync(id);
             if (category is null)
-                return NotFound();
+                return NotFound(new APIResponse<object>(404, "The category is not found.",null));
 
             var categoryDto = _mapper.Map<GetCategoryDto>(category);
-            return categoryDto;
+            return Ok(new APIResponse<GetCategoryDto>(200, "", categoryDto));
         }
 
         [HttpGet("search/{name}/{pageNumber}/{pageSize}")]
-        public async Task<ActionResult<PaginationDto<GetCategoryDto>>> GetCategories([FromRoute] string name, [FromRoute] int pageSize, [FromRoute] int pageNumber)
+        public async Task<ActionResult<APIResponse<PaginationDto<GetCategoryDto>>>> GetCategories([FromRoute] string name, [FromRoute] int pageSize, [FromRoute] int pageNumber)
         {
             var categories = await  _categoryRepository.GetCategoriesbyName(name,pageNumber, pageSize);
             var categoryDtos = _mapper.Map<PaginationDto<GetCategoryDto>>(categories);
-            return Ok(categoryDtos);
+            return Ok(new APIResponse<PaginationDto<GetCategoryDto>>(200, "", categoryDtos));
 
         }
 
@@ -58,45 +62,100 @@ namespace backend.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateCategory(int id,[FromBody] AddCategoryDto categoryDto)
         {
-            var validationResult = await _validator.ValidateAsync(categoryDto);
-            if (validationResult.Errors.Any())
+            if (!ModelState.IsValid)
             {
-                return BadRequest(validationResult);
+                var errors = ModelState
+                .Where(e => e.Value!.Errors.Count > 0)
+                .SelectMany(e => e.Value!.Errors)
+                .Select(e => e.ErrorMessage)
+                .ToList();
+                var response = new APIResponse<object>(400, string.Join("\n", errors), null);
+                return BadRequest(response);
             }
+
+            var validationResult = await _validator.ValidateAsync(categoryDto);
+            if (!validationResult.IsValid)
+            {
+                var fluentErrors = validationResult.Errors
+                    .Select(e => e.ErrorMessage)
+                    .ToList();
+
+                var response = new APIResponse<object>(400, string.Join("\n", fluentErrors), null);
+                return BadRequest(response);
+            }
+
             var existingCategory = await _categoryRepository.GetbyIdAsync(id);
             if (existingCategory == null)
             {
-                return NotFound();
+                return NotFound(new APIResponse<object>(404,"This category doesn't exists", null));
             }
 
             // Map the data from the DTO to the existing category entity
             _mapper.Map(categoryDto, existingCategory);
 
             await _categoryRepository.UpdateAsync(existingCategory);
-            return NoContent();
+            return Ok(new APIResponse<object>(200, "The category is updated successfully.", null));
         }
 
         [Authorize(Roles = "lIBRARIAN")]
         [HttpPost]
         public async Task<ActionResult<Category>> AddCategory([FromBody]AddCategoryDto categoryDto)
         {
-            var validationResult = await _validator.ValidateAsync(categoryDto);
-            if(validationResult.Errors.Any())
+            if (!ModelState.IsValid)
             {
-                return BadRequest(validationResult);
+                var errors = ModelState
+                .Where(e => e.Value!.Errors.Count > 0)
+                .SelectMany(e => e.Value!.Errors)
+                .Select(e => e.ErrorMessage)
+                .ToList();
+                var response = new APIResponse<object>(400, string.Join("\n", errors), null);
+                return BadRequest(response);
             }
-                var category = _mapper.Map<Category>(categoryDto);
+
+            var validationResult = await _validator.ValidateAsync(categoryDto);
+            if (!validationResult.IsValid)
+            {
+                var fluentErrors = validationResult.Errors
+                    .Select(e => e.ErrorMessage)
+                    .ToList();
+
+                var response = new APIResponse<object>(400, string.Join("\n", fluentErrors), null);
+                return BadRequest(response);
+
+            }
+            var category = _mapper.Map<Category>(categoryDto);
             await _categoryRepository.AddAsync(category);
-            return Ok();
+            return Ok(new APIResponse<object>(200,"The categoey added successfully." , null));
         }
 
         [Authorize(Roles = "lIBRARIAN")]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteCategory(int id)
         {
+            var existingCategory = await _categoryRepository.GetbyIdAsync(id);
+            if (existingCategory == null)
+            {
+                return NotFound(new APIResponse<object>(404, "This category doesn't exists", null));
+            }
+
             await _categoryRepository.DeleteAsync(id);
 
-            return NoContent();
+            return Ok(new APIResponse<object>(200, "The catgegory deleted successfully.", null));
+        }
+
+        [HttpGet("names")]
+        public async Task<ActionResult<APIResponse<ICollection<string>>>> GetCategoriesNames()
+        {
+            var res = await _categoryRepository.GetCategoriesNames();
+            return Ok(new APIResponse<ICollection<string>>(200,"", res));
+        }
+
+        [HttpGet("all")]
+        public async Task<ActionResult<APIResponse<ICollection<GetCategoryDto>>>> GetAll()
+        {
+            var categories = await _categoryRepository.getAllAsync();
+            var res = _mapper.Map<ICollection<GetCategoryDto>>(categories);
+            return Ok(new APIResponse<ICollection<GetCategoryDto>>(200, "", res));
         }
     }
 }
